Move movement functions to robot config and implement them
